[JarInfer] Ignore non-public classes when inferring annotations.
This improves both performance and avoids a particular error when
adding RuntimeInvisibleParameterAnnotations attributes to
package visibility instance inner classes.
