+++ jar-infer/test-android-lib-jarinfer/src/main/java/com/uber/nullaway/jarinfer/toys/Foo.java
@@ -1,0 +1,1 @@
+
