+++ jar-infer/test-android-lib-jarinfer/src/main/java/com/uber/nullaway/jarinfer/toys/unannotated/ExpectNullable.java
@@ -1,0 +1,7 @@
+package com.uber.nullaway.jarinfer.toys.unannotated;
+
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+
+@Retention(RetentionPolicy.RUNTIME)
+@interface ExpectNullable {}
