+++ jar-infer/test-java-lib-jarinfer/src/main/java/com/uber/nullaway/jarinfer/toys/ExpectNullable.java
@@ -1,0 +1,3 @@
+package com.uber.nullaway.jarinfer.toys;
+
+public class ExpectNullable {}
